+++ foxes/input/__init__.py
@@ -1,2 +1,3 @@
 from . import farm_layout
 from . import states
+from . import windio