+++ foxes/input/farm_layout/__init__.py
@@ -3,3 +3,4 @@ from .add_grid import add_grid
 from .add_from_json import add_from_json
 from .add_from_csv import add_from_csv
 from .add_from_file import add_from_file
+from .add_from_df import add_from_df