+++ foxes/output/farm_results_eval.py
@@ -359,17 +359,20 @@ class FarmResultsEval(Output):
 
     def calc_turbine_yield(
         self,
+        algo=None,
         annual=False,
         ambient=False,
         hours=None,
         delta_t=None,
-        P_unit_W=1e3,
+        P_unit_W=None,
     ):
         """
         Calculates the yield per turbine
 
         Parameters
         ----------
+        algo : foxes.core.Algorithm, optional
+            The algorithm, for P_nominal lookup
         annual : bool, optional
             Flag for returing annual results, by default False
         ambient : bool, optional
@@ -380,7 +383,8 @@ class FarmResultsEval(Output):
             The time delta step in case of time series data,
             by default automatically determined
         P_unit_W : float
-            The power unit in Watts, 1000 for kW
+            The power unit in Watts, 1000 for kW. Looked up
+            in algorithm if not given
 
         Returns
         -------
@@ -395,6 +399,16 @@ class FarmResultsEval(Output):
             var_in = FV.P
             var_out = FV.YLD
 
+        if algo is not None and P_unit_W is None:
+            P_unit_W = np.array(
+                [FC.P_UNITS[t.P_unit] for t in algo.farm_controller.turbine_types],
+                dtype=FC.DTYPE,
+            )[:, None]
+        elif algo is None and P_unit_W is not None:
+            pass
+        else:
+            raise KeyError("Expecting either 'algo' or 'P_unit_W'")
+        
         # compute yield per turbine
         if np.issubdtype(self.results[FV.STATE].dtype, np.datetime64):
             if hours is not None: